Check config before deleting it.

+++ modules/asset/group/farm_group.post_update.php
@@ -12,5 +12,7 @@ use Drupal\system\Entity\Action;
  */
 function farm_group_post_update_uninstall_asset_group_action(&$sandbox) {
   $config = Action::load('asset_group_action');
-  $config->delete();
+  if (!empty($config)) {
+    $config->delete();
+  }
 }
